requirements.txt
streamlit>=1.44,<2
pandas>=2.2,<3
numpy>=1.26,<3
plotly>=6.0,<7
kaleido>=0.2.1,<1
scikit-learn>=1.5,<2
scipy>=1.13,<2
psycopg[binary]>=3.2,<4
openpyxl>=3.1,<4
openai>=1.30,<2
reportlab>=4.2,<5
python-docx>=1.1,<2
